requirements.txt
streamlit==1.28.0
langchain==0.1.0
langchain-ollama==0.1.0
beautifulsoup4==4.12.0
lxml==4.9.0
html5lib==1.1
python-dotenv==1.0.0
validators==0.22.0
pandas==2.1.0
openpyxl==3.1.0
requests==2.31.0